requirements.txt
absl-py==2.0.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
configparser==6.0.0
crashtest==0.4.1
cryptography==41.0.5
distlib==0.3.7
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.13.1
google-api-core==2.13.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-access-context-manager==0.1.16
google-cloud-asset==3.20.0
google-cloud-org-policy==1.8.2
google-cloud-os-config==1.15.3
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
more-itertools==10.1.0
msgpack==1.0.7
mysql-connector==2.2.9
packaging==23.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
proto-plus==1.22.3
protobuf==4.25.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyMySQL==1.1.0
pyproject_hooks==1.0.0
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==12.6.0
rsa==4.9
SecretStorage==3.3.3
shellingham
shutils==0.1.0
tomlkit==0.12.2
trove-classifiers==2023.11.7
typer==0.7.0
typer-cli==0.0.13
urllib3==2.0.7
virtualenv==20.24.6
